implement hash funcs
Learnt some basic hash functions and their implementations.
Hashes are a cool concept and its nice how they help to map large
string to smaller integer values that are useful. There was a bit of
mathematics and plenty bitwise arithmetic, but they are pretty cool once
you get the hang of it